requirements.txt
Django~=4.2.6
djangorestframework~=3.14.0
django-cors-headers~=4.3.0
whitenoise~=6.6.0
psycopg2-binary~=2.9.9
redis~=4.5.4
django-request-logging~=0.7.5
uvicorn~=0.23.2
Unidecode~=1.3.7
Pillow~=10.1.0
djangorestframework_simplejwt~=5.3.0
PyJWT~=2.8.0
requests~=2.31.0
attrs~=21.4.0
python-dotenv~=1.0.1
typesense
minio~=7.2.8
celery~=5.4.0
django-otp~=1.5.4
django_redis~=4.11.0
